Fixed object storage command line option

+++ objectstore/server/src/main/java/micheal65536/vienna/objectstore/server/Main.java
@@ -40,7 +40,7 @@ public class Main
 		{
 			commandLine = new DefaultParser().parse(options, args);
 			dataDir = commandLine.getOptionValue("dataDir");
-			port = commandLine.hasOption("port") ? (int) commandLine.getParsedOptionValue("port") : 5396;
+			port = commandLine.hasOption("port") ? (int) (long) commandLine.getParsedOptionValue("port") : 5396;
 		}
 		catch (ParseException exception)
 		{
